requirements.txt
streamlit==1.19.0
sqlalchemy==2.0.10
werkzeug==2.3.5
pandas==2.0.2
PyPDF2==3.0.1
python-docx==0.8.11
cryptography==40.0.2
